move wikipedia_parser
